Login example
This is only a template, a lot of stuff here to be removed or changed, issues I am aware of:

The db doesn't straight away connect, so I need await function or something,

The login is not correctly styled.

the get request need to be changed to post with the content headers etc

this is good start at least.